Rework drill_uaf_callback_rop_smap.c to allow finishing gracefully
Currently this PoC doesn't finish properly if you type `exit` in the
root shell, the program hangs.

Let's do a cool experiment and reorganize forking:
1) Execute the ROP chain in a child instead of the parent process.
2) Create a separate child for provoking a core dump.
3) Call `pause()` in the parent process to wait for a signal from
the core dump helper.
4) In the core dump helper, determine a PID of the parent process and
send SIGTERM to it after the root shell is finished.

That rework allows to finish the PoC gracefully and return to
the initial shell. Nice.

However, we still have one child performing `msleep()` in the
kernelspace at the end of the ROP chain. It may crash the kernel, and it is
the main source of instability of this prototype.